Allow custom embeddings based on the task Fixes #18
