+++ src/seb/seb_models/__init__.py
@@ -1,3 +1,4 @@
+from .e5_models import *
 from .fairseq_models import *
 from .hf_models import *
 from .openai_models import *
