+++ src/seb/seb_models/hf_models.py
@@ -233,98 +233,6 @@ def create_xlm_roberta_base() -> SebModel:
     )
 
 
-# English
-@models.register("intfloat/e5-small")
-def create_e5_small() -> SebModel:
-    hf_name = "intfloat/e5-small"
-    meta = ModelMeta(
-        name=hf_name.split("/")[-1],
-        huggingface_name=hf_name,
-        reference=f"https://huggingface.co/{hf_name}",
-        languages=["en"],
-    )
-    return SebModel(
-        loader=partial(get_sentence_transformer, model_name=hf_name),  # type: ignore
-        meta=meta,
-    )
-
-
-@models.register("intfloat/e5-base")
-def create_e5_base() -> SebModel:
-    hf_name = "intfloat/e5-base"
-    meta = ModelMeta(
-        name=hf_name.split("/")[-1],
-        huggingface_name=hf_name,
-        reference=f"https://huggingface.co/{hf_name}",
-        languages=["en"],
-    )
-    return SebModel(
-        loader=partial(get_sentence_transformer, model_name=hf_name),  # type: ignore
-        meta=meta,
-    )
-
-
-@models.register("intfloat/e5-large")
-def create_e5_large() -> SebModel:
-    hf_name = "intfloat/e5-large"
-    meta = ModelMeta(
-        name=hf_name.split("/")[-1],
-        huggingface_name=hf_name,
-        reference=f"https://huggingface.co/{hf_name}",
-        languages=["en"],
-    )
-    return SebModel(
-        loader=partial(get_sentence_transformer, model_name=hf_name),  # type: ignore
-        meta=meta,
-    )
-
-
-# Multilingual
-@models.register("intfloat/multilingual-e5-small")
-def create_multilingual_e5_small() -> SebModel:
-    hf_name = "intfloat/multilingual-e5-small"
-    meta = ModelMeta(
-        name=hf_name.split("/")[-1],
-        huggingface_name=hf_name,
-        reference=f"https://huggingface.co/{hf_name}",
-        languages=[],
-    )
-    return SebModel(
-        loader=partial(get_sentence_transformer, model_name=hf_name),  # type: ignore
-        meta=meta,
-    )
-
-
-@models.register("intfloat/multilingual-e5-base")
-def create_multilingual_e5_base() -> SebModel:
-    hf_name = "intfloat/multilingual-e5-base"
-    meta = ModelMeta(
-        name=hf_name.split("/")[-1],
-        huggingface_name=hf_name,
-        reference=f"https://huggingface.co/{hf_name}",
-        languages=[],
-    )
-    return SebModel(
-        loader=partial(get_sentence_transformer, model_name=hf_name),  # type: ignore
-        meta=meta,
-    )
-
-
-@models.register("intfloat/multilingual-e5-large")
-def create_multilingual_e5_large() -> SebModel:
-    hf_name = "intfloat/multilingual-e5-large"
-    meta = ModelMeta(
-        name=hf_name.split("/")[-1],
-        huggingface_name=hf_name,
-        reference=f"https://huggingface.co/{hf_name}",
-        languages=[],
-    )
-    return SebModel(
-        loader=partial(get_sentence_transformer, model_name=hf_name),  # type: ignore
-        meta=meta,
-    )
-
-
 # Scandinavian sentence encoders
 @models.register("KennethEnevoldsen/dfm-sentence-encoder-large-1")
 def create_dfm_sentence_encoder_large() -> SebModel:
